Refact : refact lightning module
- eyes, skin module 분리
