+++ src/module/__init__.py
@@ -1,0 +1,4 @@
+from .classification import ClassificationModel
+from .segmentation import SegmentationModel
+
+__all__ = ["ClassificationModel", "SegmentationModel"]
